requirements.txt
bcrypt==4.1.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.3
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
ptyprocess==0.6.0
Pygments==2.2.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.1.7
Werkzeug==3.0.1
WTForms==3.1.1
zope.interface==6.1
